requirements.txt
Django==3.1.7
djangorestframework==3.12.2
pylint==2.7.2
autopep8==1.5.5
bcrypt==3.2.0
pillow==8.1.1
pandas==1.2.3